Artificial Intelligence Performing Image Classification with Transfer Learning
In this project we will make image classification using Transfer Learning. For this purpose we will use the InceptionResNetV2 model which is trained on the imageNet data set. Lets use a submarine image for classification. I found this image in Internet, this image is not included in imageNet dataset which was used for training InceptionResNetV2.

In transfer learning we use a model that has been previously trained on a dataset and contains the weights and biases that represent the features of whichever dataset it was trained on. 

Inception and ResNet have been among the best image recognition performance models in recent years, with very good performance at a relatively low computational cost. Inception and ResNet combines the Inception architecture, with residual connections.